Restore tab selection on restart.
This change add configuration to the application so that when the application is restarted, the previously selected tab is displayed.

Add a “restoration id” to the tab view controller and top level view controllers.
Remove from the app delegate the selection of the shift tab.
Add to the app delegate the methods shouldSaveApplicationState and shouldRestoreApplicationState to opt in to the state restoration mechanism.
Update the frank navigation test to verify