Prepare for running on CI Server
The CI server requires the “schemes” to be shared - not related to a user.

The strategy with CocoaPods is to check in the source code of the pods but not the post “pod install” state. The file Podfile.lock tracks the specific “pods” used. The file Pods/Manifest.lock represents the installed state of the Pods directory. When the “Check Pods Manifest.lock” build phase script is run, it now checks to see if Podfile.lock matches Manifest.lock and runs “pod install” if required.

This change:
Removes Manifest.lock. This is generated by “pod install”.
Change the Pod project schemes to be shared so that the CI server can run them.
Changes the “Check Pods Manifest.lock” script to automatically run “pod install” when required.
Adds Pods-ppreminder.xcconfig to the ppreminderer build to ensure the settings are available for the build.