Add scheduler time simulation and settings controls.
This change adds to the scheduler the concept of scheduler time which can be used to test the scheduling of the application.
The scheduler time maintained by a one second timer. Each time the scheduler time is updates, a notification is posted.
The settings tab allows the time to be set and a “warp” factor which can be used to speed up the “clock”.

PPRClockTabBarItem has been added to display the scheduler time.
PPRSettingViewController has been modified to replace the “Settings” tab bar label with a ClockTabBarItem and register the clock display for update notification.
PPRScheduler is modified to manage the schedulerTime and provides routines to calculate the equivalent wall time clock value for a schedulerTime value.
PPRAppDelegate has been modified to start the scheduler timer and to post notifications when the schedulerTime has changed. Experimental NoticeView messages have been removed and a more generic display of notifications provided.
Main.storyboard has had controls added to control the schedulerTime.