Improve handling of schedulerTime
SchedulerTime now is maintained by capturing the wall clock time and the scheduler time at each tick and maintaining the relationship between them rather then “counting” the timer ticks.
When the scheduler time is manually set, the UILocalNotifications are rescheduled.
Care is taken to try to reduce coupling so that PPRScheduler and PPRNotificationManager should not need to know anything about actions. This is not yet completely there.

PPRStickyMessageNoticeView
- remove unused variable

PPRSettingsNavigationControllerViewController.
- Change the name of the notification for scheduler time updates

PPRSettingsViewController
- Change how the schedulerTime is set and maintained

PPRActionManager
- make public “getActionById”
- notify changes to action state

PPRClientManager
- Improve comments

PPRFacilityManager
- Improve comments

PPRNotificationManager
- If the schedule time is reset then reschedule notifications
- Provide a means of removing notifications

PPRScheduler
- Change the handing of scheduled time
- Provide a means to reschedule items.
- Factor out scheduling items and sorting items
- timer tick handling captures the scheduler time and the wall clock time and preserves the relationship between them.